[move-bytecode-util] Move ModuleCache to `move-bytecode-util`
Closes: #9577
