+++ language/move-binary-format/src/lib.rs
@@ -20,7 +20,6 @@ pub mod deserializer;
 pub mod file_format;
 pub mod file_format_common;
 pub mod internals;
-pub mod layout;
 pub mod normalized;
 #[cfg(any(test, feature = "fuzzing"))]
 pub mod proptest_types;
