+++ language/tools/read-write-set/dynamic/Cargo.toml
@@ -13,5 +13,6 @@ edition = "2018"
 anyhow = "1.0.38"
 diem-workspace-hack = { path = "../../../../common/workspace-hack" }
 move-binary-format = { path = "../../../move-binary-format" }
+move-bytecode-utils = { path = "../../move-bytecode-utils" }
 move-core-types = { path = "../../../move-core/types" }
 read-write-set-types = { path = "../types" }
